Revision 4: Move AppDebugInfo to layout components
