+++ apps/opik-frontend/src/components/layout/AppDebugInfo/AppDebugInfo.tsx
