+++ apps/opik-frontend/src/components/layout/PartialPageLayout/PartialPageLayout.tsx
@@ -3,7 +3,7 @@ import useAppStore from "@/store/AppStore";
 import usePluginsStore from "@/store/PluginsStore";
 import { Link, Outlet } from "@tanstack/react-router";
 import Logo from "@/components/layout/Logo/Logo";
-import AppDebugInfo from "@/components/shared/AppDebugInfo/AppDebugInfo";
+import AppDebugInfo from "@/components/layout/AppDebugInfo/AppDebugInfo";
 import ThemeToggle from "@/components/layout/ThemeToggle/ThemeToggle";
 
 export const PartialPageLayout = ({
